rlc-orchestrator/requirements.txt
# RLC Orchestrator - Python Dependencies
# ========================================
# Install with: pip install -r requirements.txt

# Core dependencies
sqlalchemy>=2.0.0          # Database ORM
python-dotenv>=1.0.0       # Environment variable management

# HTTP requests (for API calls)
requests>=2.31.0           # Synchronous HTTP
httpx>=0.25.0              # Async HTTP (for future use)

# Date/Time handling
python-dateutil>=2.8.0     # Date parsing and manipulation
croniter>=1.4.0            # Cron expression parsing

# Logging and monitoring
structlog>=23.0.0          # Structured logging (optional upgrade from stdlib)

# Email handling (Phase 2)
# Uncomment when you implement email integration
# imapclient>=2.3.0        # IMAP client for receiving email
# python-email-validator>=2.0.0  # Email address validation

# AI APIs (Phase 3)
# Uncomment when you implement AI integration
# anthropic>=0.18.0        # Claude API
# openai>=1.0.0            # OpenAI API (optional alternative)

# Local LLM (Optional - for hybrid approach)
# Uncomment if you want to run local models
# ollama>=0.1.0            # Ollama Python client

# Development and testing
pytest>=7.4.0              # Testing framework
pytest-asyncio>=0.21.0     # Async test support
black>=23.0.0              # Code formatting
ruff>=0.1.0                # Linting

# Production (uncomment when deploying)
# gunicorn>=21.0.0         # WSGI server (if adding web interface)
# uvicorn>=0.24.0          # ASGI server (if adding async web interface)
